Move the kwdagger machinery into a private module
GenericPipelineProcessor and KWDaggerProcessor are the two adapters
between an evaluation card and a kwdagger pipeline. They sit in the
middle of evaluation.py, between EvaluationCard and EvaluationTask,
where they crowd out the card logic and make every change to either one
look like a change to the other.

Move both to magnet/_kwdagger.py verbatim, along with the kwdagger
imports they need. The only edits are the two doctest import lines,
which now name the new module. evaluation.py imports both names, so
`from magnet.evaluation import KWDaggerProcessor` still works.
