+++ magnet/_kwdagger.py
@@ -1,0 +1,249 @@
+"""
+Bridge between an evaluation card and a kwdagger pipeline.
+
+A card can specify its pipeline in two ways: a ``pipeline:`` block, which
+:class:`GenericPipelineProcessor` turns into a kwdagger DAG, or a ``kwdagger:``
+block, which :class:`KWDaggerProcessor` passes through as a schedule spec.
+"""
+import json
+from typing import Any, Dict, List, Tuple
+
+import ubelt as ub
+from kwdagger import Pipeline, ProcessNode
+from kwdagger.schedule import ScheduleEvaluationConfig, build_schedule
+
+
+class GenericPipelineProcessor:
+    """
+    Handler for yaml-based pipeline specification
+
+    NOTE:
+        *possibly merge with KWDaggerProcessor*
+
+    Example:
+        >>> from magnet._kwdagger import GenericPipelineProcessor
+        >>> import kwutil
+        >>> # Example snippet of an Evaluation Card
+        >>> example_cfg = kwutil.Yaml.coerce(
+            '''
+            pipeline:
+              predict_node:
+                executable: python -m magnet.examples.llama_consistency.llama_predict
+                algo_params:
+                  base_model: ["meta/llama-2-13b", "meta/llama-2-70b"]
+                  comp_model: ["meta/llama-2-7b", "meta/llama-3-70b"]
+                out_paths:
+                  results_fpath: ./llama_results.json
+            ''')
+        >>> root_dpath = "."
+        >>> pipeline_def = example_cfg['pipeline']
+        >>> pipeline = GenericPipelineProcessor(pipeline_def, root_dpath)
+        >>> #
+        >>> # Construct One Node Pipeline
+        >>> pipeline.define_kwdagger()
+        ...
+        >>> pipeline.dag.print_graphs()
+
+        Process Graph
+        ╙── predict_node
+
+        IO Graph
+        ╙── predict_node
+            ╽
+            results_fpath
+
+        >>> for attr in ['name', 'executable', 'algo_params', 'out_paths']:
+        >>>    print(getattr(pipeline.dag.node_dict['predict_node'], attr))
+        predict_node
+        python -m magnet.examples.llama_consistency.llama_predict
+        ['base_model', 'comp_model']
+        {'results_fpath': './llama_results.json'}
+        >>> #
+        >>> # Parameters matrix
+        >>> pipeline.matrix
+        {'predict_node.base_model': ['meta/llama-2-13b', 'meta/llama-2-70b'],
+        'predict_node.comp_model': ['meta/llama-2-7b', 'meta/llama-3-70b']}
+    """
+
+    def __init__(
+        self, pipeline_def: Dict[str, Any], root_dpath: ub.Path
+    ) -> None:
+        self.pipeline = pipeline_def
+        self.root_dpath = root_dpath
+        self.dag = None
+        self.matrix = None
+        self.symbols = {}
+
+    def define_kwdagger(self) -> None:
+        """
+        Construct kwdagger pipeline programmatically
+
+        *only verified for one-stage pipeline, needs 'connector' handling*
+        """
+        nodes = {}
+
+        for node_name in self.pipeline:
+            # collect nodes
+            node_params = self.pipeline[node_name]
+
+            # FIXME: should update matrix for full pipeline
+            node_params, self.matrix = self._parse_params(
+                node_name, node_params
+            )
+
+            node = ProcessNode(name=node_name, **node_params)
+            nodes[node_name] = node
+
+        self.dag = Pipeline(list(nodes.values()))
+        self.dag.build_nx_graphs()
+
+    def dispatch(
+        self, backend: str = 'serial', skip_existing: bool = True, **kwargs: Any
+    ) -> None:
+        self.define_kwdagger()
+
+        kwdagger_params = {'pipeline': self.dag, 'matrix': self.matrix}
+
+        kwd_config = ScheduleEvaluationConfig(
+            params=kwdagger_params,  # includes pipeline and additional params
+            root_dpath=self.root_dpath,
+            backend=backend,
+            skip_existing=skip_existing,
+            run=True,
+        )
+
+        dag, queue = build_schedule(kwd_config)
+
+    def collect_symbols(self) -> Dict[str, Any]:
+        """
+        Collect results (Evaluation Card 'symbols') in place of 'load_result' in the ProcessNode definition
+        """
+        if not self.symbols:
+            self.dispatch()
+
+        # Glob all results json (only one node in pipeline)
+        paths = self.root_dpath.glob(
+            f'**/{self.dag.node_dict[next(iter(self.dag.node_dict))].out_paths["results_fpath"]}'
+        )
+
+        for symbol_resolution in paths:
+            symbols = json.load(open(symbol_resolution, 'r'))
+            parent_dir = symbol_resolution.parent.stem
+            if 'result' in symbols:
+                # assume all fields exist
+                for symbol in symbols['result']:
+                    # record all sweeps
+                    if parent_dir not in self.symbols:
+                        self.symbols[parent_dir] = {}
+
+                    self.symbols[parent_dir][symbol] = {
+                        'value': symbols['result'][symbol]
+                    }
+
+        return self.symbols
+
+    def _parse_params(
+        self, node_name: str, node_cfg: Dict[str, Any]
+    ) -> Tuple[Dict[str, Any], Dict[str, Any]]:
+        """
+        Parse sweepable parameters from definition
+        """
+        matrix = {}
+        for k in node_cfg:
+            if isinstance(node_cfg[k], dict) and '_params' in k:
+                # TODO: Construct a more robust validator
+                for param, v in node_cfg[k].items():
+                    matrix[f'{node_name}.{param}'] = v
+                # decompose yaml
+                node_cfg[k] = list(node_cfg[k].keys())
+        return node_cfg, matrix
+
+
+class KWDaggerProcessor:
+    """
+    Handler for full kwdagger pipeline specification
+
+    Example
+        >>> from magnet._kwdagger import KWDaggerProcessor
+        >>> from kwdagger.schedule import ScheduleEvaluationConfig, build_schedule
+        >>> import kwutil
+        >>> # Example snippet of an Evaluation Card (related to GenericPipelineProcessor example)
+        >>> example_cfg = kwutil.Yaml.coerce(
+            '''
+            kwdagger:
+              pipeline: magnet.examples.llama_consistency.pipelines.llama_pipeline()
+              matrix:
+                llama_predict.base_model: ["meta/llama-2-13b", "meta/llama-2-70b"]
+                llama_predict.comp_model:  ["meta/llama-2-7b", "meta/llama-3-70b"]
+            ''')
+        >>> root_dpath = "."
+        >>> kwdagger_def = example_cfg['kwdagger']
+        >>> pipeline = KWDaggerProcessor(kwdagger_def, root_dpath)
+        >>> #
+        >>> # Construct Two Node Pipeline (llama_predict -> claim)
+        >>> kwdagger_spec = ScheduleEvaluationConfig(params=pipeline.spec, run=False)
+        >>> dag, queue = build_schedule(kwdagger_spec)
+        ...
+        >>> dag.print_graphs()
+
+        Process Graph
+        ╙── llama_predict
+            ╽
+            claim_eval
+
+        IO Graph
+        ╙── llama_predict
+            ╽
+            results_fpath
+            ╽
+            symbols_fpath
+            ╽
+            claim_eval
+            ╽
+            verdict_fpath
+
+        >>> #
+        >>> # Parameters matrix
+        >>> pipeline.spec['matrix']
+        {'llama_predict.base_model': ['meta/llama-2-13b', 'meta/llama-2-70b'],
+        'llama_predict.comp_model': ['meta/llama-2-7b', 'meta/llama-3-70b']}
+    """
+
+    def __init__(
+        self, pipeline_def: Dict[str, Any], root_dpath: ub.Path
+    ) -> None:
+        self.spec = pipeline_def
+        self.root_dpath = root_dpath
+        self.results = []
+        self.symbols = []
+
+    def dispatch(
+        self, backend: str = 'serial', skip_existing: bool = True, **kwargs: Any
+    ) -> None:
+        kwd_config = ScheduleEvaluationConfig(
+            params=self.spec,  # includes pipeline and additional params
+            root_dpath=self.root_dpath,
+            backend=backend,
+            skip_existing=skip_existing,
+            run=True,
+            **kwargs,
+        )
+
+        self.dag, queue = build_schedule(kwd_config)
+
+    def collect_results(self) -> Tuple[List[str], List[Any]]:
+        if not self.results:
+            self.dispatch()
+
+        # Glob all Claim node json files recursively
+        paths = self.root_dpath.glob('**/verdict.json')
+
+        # Assumes {result: {status: value}} output format
+        for claim_json in paths:
+            claim_result = json.load(open(claim_json, 'r'))
+            if 'result' in claim_result and 'status' in claim_result['result']:
+                self.results.append(claim_result['result']['status'])
+            if 'result' in claim_result and 'symbols' in claim_result['result']:
+                self.symbols.append(claim_result['result']['symbols'])
+
+        return self.results, self.symbols
